Separated variableAttribute class from variableAttributeLoader files
