+++ include/variableAttributeLoader.h
@@ -1,112 +1,14 @@
-// Class to hold the variable attributes that will be passed to a
-// userInputParameters object
+// Class to manage the variable attributes that the user specifies
 #ifndef VARIABLEATTRIBUTELOADER_H
 #define VARIABLEATTRIBUTELOADER_H
 
-#include "model_variables.h"
 #include "varTypeEnums.h"
+#include "variableAttributes.h"
 
-#include <set>
+#include <map>
 #include <string>
-#include <vector>
 
 using EvalFlags = dealii::EvaluationFlags::EvaluationFlags;
-
-struct variableAttributes
-{
-  // Variable inputs (v2.0)
-  std::string name                  = "";
-  fieldType   var_type              = UNDEFINED_FIELD;
-  PDEType     eq_type               = UNDEFINED_PDE;
-  bool        need_value_nucleation = false;
-  bool        nucleating_variable   = false;
-  bool        is_pp                 = false;
-  bool        is_nonlinear          = false;
-  bool        calc_integral         = false;
-  bool        output_integral       = false;
-
-  std::set<std::string> dependencies_value_RHS;
-  std::set<std::string> dependencies_gradient_RHS;
-  std::set<std::string> dependencies_RHS;
-  std::set<std::string> dependencies_value_LHS;
-  std::set<std::string> dependencies_gradient_LHS;
-  std::set<std::string> dependencies_LHS;
-  std::set<std::string> dependencies_value_PP;
-  std::set<std::string> dependencies_gradient_PP;
-  std::set<std::string> dependencies_PP;
-
-  std::set<std::string> dependency_set;
-
-  EvalFlags eval_flags_explicit_RHS    = dealii::EvaluationFlags::nothing;
-  EvalFlags eval_flags_nonexplicit_RHS = dealii::EvaluationFlags::nothing;
-  EvalFlags eval_flags_nonexplicit_LHS = dealii::EvaluationFlags::nothing;
-
-  EvalFlags eval_flags_change_nonexplicit_LHS = dealii::EvaluationFlags::nothing;
-
-  EvalFlags eval_flags_residual_explicit_RHS    = dealii::EvaluationFlags::nothing;
-  EvalFlags eval_flags_residual_nonexplicit_RHS = dealii::EvaluationFlags::nothing;
-  EvalFlags eval_flags_residual_nonexplicit_LHS = dealii::EvaluationFlags::nothing;
-
-  EvalFlags eval_flags_postprocess          = dealii::EvaluationFlags::nothing;
-  EvalFlags eval_flags_residual_postprocess = dealii::EvaluationFlags::nothing;
-
-  void
-  format_dependencies();
-  void
-  parse_dependencies(std::map<uint, variableAttributes> &other_var_attributes);
-
-  void
-  parse_residual_dependencies();
-
-  std::set<EvalFlags *>
-  eval_flags_for_eq_type(const variableAttributes &other_variable);
-
-  /*   std::set<EvalFlags *>
-    residual_flags_for_eq_type(PDEType other_eq_type)
-    {
-      if (other_eq_type == EXPLICIT_TIME_DEPENDENT)
-        {
-          return {&eval_flags_residual_explicit_RHS};
-        }
-      if (other_eq_type == AUXILIARY)
-        {
-          return {&eval_flags_residual_nonexplicit_RHS};
-        }
-      if (other_eq_type == IMPLICIT_TIME_DEPENDENT || other_eq_type == TIME_INDEPENDENT)
-        {
-          return {&eval_flags_residual_nonexplicit_RHS,
-                  &eval_flags_residual_nonexplicit_LHS};
-        }
-      return {};
-    } */
-
-  /* std::set<std::pair<std::set<std::string> *, std::set<std::string> *>>
-    dep_set_for_eq_type(PDEType other_eq_type)
-    {
-      if (other_eq_type == EXPLICIT_TIME_DEPENDENT)
-        {
-          return dependencies_RHS;
-          // return {
-          //   {&dependencies_value_RHS, &dependencies_gradient_RHS}
-          // };
-        }
-      if (other_eq_type == AUXILIARY)
-        {
-          return dependencies_RHS;
-          // return {
-          //   {&dependencies_value_RHS, &dependencies_gradient_RHS}
-          // };
-        }
-      if (other_eq_type == IMPLICIT_TIME_DEPENDENT || other_eq_type == TIME_INDEPENDENT)
-        {
-          return dependencies_RHS;
-          // return {
-          //   {&dependencies_value_RHS, &dependencies_gradient_RHS}
-          // };
-        }
-      return {};
-    } */
-};
 
 class variableAttributeLoader
 {
@@ -163,7 +65,6 @@ public:
   strip_whitespace(const std::string &text);
 
   // Members
-
   std::map<uint, variableAttributes>  attributes;
   std::map<uint, variableAttributes>  pp_attributes;
   std::map<uint, variableAttributes> *relevant_attributes = nullptr;
